requirements.txt
Flask
docker
six
gunicorn
flask-cors
hurry.filesize
psutil
gevent
flask-login
python-pam
pyjwt
